feat: LinkifyCompat on TextView after text is set
The `autoLinkAll` extension function now uses `LinkifyCompat.addLinks` after the text is set on the TextView.
This ensures that links are correctly added.
The change does not use `autoLinkMask` anymore since the original method worked better on older api.
It adds links for `WEB_URLS`, `PHONE_NUMBERS`, and `EMAIL_ADDRESSES`.