Refactor: Introduce `onBoundsChange` property and deprecate function
This commit introduces a new `onBoundsChange` property to `View` that returns a `CSHasChange<Unit>`. This provides a more modern and idiomatic way to observe bounds changes compared to the previous `onBoundsChange(function: Func)` function, which is now deprecated.

- Added a new `val View.onBoundsChange: CSHasChange<Unit>` extension property.
- The `onChange` method of this property can be used to register a lambda that will be invoked when the view's bounds change.
- The existing `View.onBoundsChange(function: Func): CSRegistration` function has been marked as `@Deprecated`.
- The `hasSize(parent: CSHasRegistrations?)` function was updated to use the new `onBoundsChange` property.

Additionally, the commit includes minor code reformatting for improved readability in `View+CSRegistration.kt`.