Feat: Add `onLoad` event to `CSGridItemView`
This commit introduces a new public `onLoad` event to the `CSGridItemView`.

The event is triggered at the end of the `load` function, passing the loaded data. This provides a more flexible way for external observers to react when the item view is loaded with data, complementing the existing `onLoad` override function.