feat: Add `normal` alpha parameter to `disabledByAlphaIf`
This commit enhances the `disabledByAlphaIf` extension functions for `View` by adding a `normal` alpha parameter. This allows specifying a custom alpha value for the view when it is in its normal (not disabled) state, defaulting to `1F`.

The new `normal` parameter has been added to both overloads of the function, providing more flexibility in controlling view appearance based on a property's state.